# perf(es/minifier): Merge `MultiReplacer` calls