--- a/crates/swc_ecma_minifier/src/compress/optimize/mod.rs
+++ b/crates/swc_ecma_minifier/src/compress/optimize/mod.rs
@@ -262,22 +262,14 @@ impl Vars {
         N: for<'aa> VisitMutWith<MultiReplacer<'aa>>,
     {
         let mut changed = false;
-        if !self.simple_functions.is_empty() {
-            n.visit_mut_with(&mut MultiReplacer::new(
-                &mut self.simple_functions,
-                true,
-                MultiReplacerMode::OnlyCallee,
-                &mut changed,
-            ));
-        }
 
-        if !self.vars_for_inlining.is_empty() {
-            n.visit_mut_with(&mut MultiReplacer::new(
-                &mut self.vars_for_inlining,
-                false,
-                MultiReplacerMode::Normal,
-                &mut changed,
-            ));
+        if !self.simple_functions.is_empty() || !self.vars_for_inlining.is_empty() {
+            n.visit_mut_with(&mut MultiReplacer {
+                vars: &mut self.vars_for_inlining,
+                simple_functions: &self.simple_functions,
+                changed: false,
+                worked: &mut changed,
+            });
         }
 
         changed