--- a/crates/swc_ecma_minifier/src/compress/optimize/util.rs
+++ b/crates/swc_ecma_minifier/src/compress/optimize/util.rs
@@ -191,16 +191,14 @@ impl VisitMut for MultiReplacer<'_> {
     fn visit_mut_callee(&mut self, e: &mut Callee) {
         e.visit_mut_children_with(self);
 
-        if matches!(self.mode, MultiReplacerMode::OnlyCallee) {
-            if let Callee::Expr(e) = e {
-                if let Expr::Ident(i) = &**e {
-                    if let Some(new) = self.var(&i.to_id()) {
-                        debug!("multi-replacer: Replaced `{}`", i);
-                        *self.worked = true;
-                        self.changed = true;
+        if let Callee::Expr(e) = e {
+            if let Expr::Ident(i) = &**e {
+                if let Some(new) = self.var(&i.to_id(), MultiReplacerMode::OnlyCallee) {
+                    debug!("multi-replacer: Replaced `{}`", i);
+                    *self.worked = true;
+                    self.changed = true;
 
-                        **e = *new;
-                    }
+                    **e = *new;
                 }
             }
         }
@@ -216,15 +214,13 @@ impl VisitMut for MultiReplacer<'_> {
             return;
         }
 
-        if matches!(self.mode, MultiReplacerMode::Normal) {
-            if let Expr::Ident(i) = e {
-                if let Some(new) = self.var(&i.to_id()) {
-                    debug!("multi-replacer: Replaced `{}`", i);
-                    *self.worked = true;
-                    self.changed = true;
+        if let Expr::Ident(i) = e {
+            if let Some(new) = self.var(&i.to_id(), MultiReplacerMode::Normal) {
+                debug!("multi-replacer: Replaced `{}`", i);
+                *self.worked = true;
+                self.changed = true;
 
-                    *e = *new;
-                }
+                *e = *new;
             }
         }
     }
@@ -251,21 +247,19 @@ impl VisitMut for MultiReplacer<'_> {
     fn visit_mut_prop(&mut self, p: &mut Prop) {
         p.visit_mut_children_with(self);
 
-        if matches!(self.mode, MultiReplacerMode::Normal) {
-            if let Prop::Shorthand(i) = p {
-                if let Some(value) = self.var(&i.to_id()) {
-                    debug!("multi-replacer: Replaced `{}` as shorthand", i);
-                    *self.worked = true;
-                    self.changed = true;
+        if let Prop::Shorthand(i) = p {
+            if let Some(value) = self.var(&i.to_id(), MultiReplacerMode::Normal) {
+                debug!("multi-replacer: Replaced `{}` as shorthand", i);
+                *self.worked = true;
+                self.changed = true;
 
-                    *p = Prop::KeyValue(KeyValueProp {
-                        key: PropName::Ident(Ident::new(
-                            i.sym.clone(),
-                            i.span.with_ctxt(Default::default()),
-                        )),
-                        value,
-                    });
-                }
+                *p = Prop::KeyValue(KeyValueProp {
+                    key: PropName::Ident(Ident::new(
+                        i.sym.clone(),
+                        i.span.with_ctxt(Default::default()),
+                    )),
+                    value,
+                });
             }
         }
     }