--- a/crates/swc_ecma_minifier/src/compress/optimize/util.rs
+++ b/crates/swc_ecma_minifier/src/compress/optimize/util.rs
@@ -229,7 +229,7 @@ impl VisitMut for MultiReplacer<'_> {
     fn visit_mut_module_items(&mut self, items: &mut Vec<ModuleItem>) {
         loop {
             self.changed = false;
-            if self.vars.is_empty() {
+            if self.vars.is_empty() && self.simple_functions.is_empty() {
                 break;
             }
             items.visit_mut_children_with(self);