--- a/crates/swc_ecma_minifier/src/compress/optimize/util.rs
+++ b/crates/swc_ecma_minifier/src/compress/optimize/util.rs
@@ -206,12 +206,12 @@ impl VisitMut for MultiReplacer<'_> {
     }
 
     fn visit_mut_expr(&mut self, e: &mut Expr) {
-        if self.vars.is_empty() {
+        if self.vars.is_empty() && self.simple_functions.is_empty() {
             return;
         }
         e.visit_mut_children_with(self);
 
-        if self.vars.is_empty() {
+        if self.vars.is_empty() && self.simple_functions.is_empty() {
             return;
         }