@kdy1 2022-09-06
Closing as this actually regresses performance